requirements.txt
python-dateutil==2.8.2
typing-extensions>=4.11
openai>=1.107.0
requests==2.32.3
python-dotenv==1.0.1
